feat: changes query

See merge request keibiengine/keibi-engine!497

+++ pkg/inventory/api/dto.go
@@ -481,7 +481,7 @@ type FilterInsight struct {
 	FilterID      string      `json:"filterID"`
 	CloudProvider source.Type `json:"cloudProvider"`
 	Name          string      `json:"name"`
-	Value         int64       `json:"value"`
+	Value         int         `json:"value"`
 	ValueChange   *float64    `json:"valueChange,omitempty"`
 }
